test(installer): drain PTY output before reaping
